refactor : store 도메인 분리
