+++ blackWhiteDeliveryDriver/src/main/java/com/sparta/blackwhitedeliverydriver/store/dto/StoreRequestDto.java
@@ -1,4 +1,4 @@
-package com.sparta.blackwhitedeliverydriver.dto;
+package com.sparta.blackwhitedeliverydriver.store.dto;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Pattern;
@@ -8,9 +8,7 @@ import lombok.Setter;
 
 @Getter
 @Setter
-public class StoreByMasterRequestDto {
-    @NotBlank
-    private String username;
+public class StoreRequestDto {
     @NotBlank
     private String storeName;
     @NotBlank
