+++ blackWhiteDeliveryDriver/src/main/java/com/sparta/blackwhitedeliverydriver/store/entity/StoreCategory.java
@@ -1,5 +1,6 @@
-package com.sparta.blackwhitedeliverydriver.entity;
+package com.sparta.blackwhitedeliverydriver.store.entity;
 
+import com.sparta.blackwhitedeliverydriver.category.entity.Category;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
