+++ blackWhiteDeliveryDriver/src/main/java/com/sparta/blackwhitedeliverydriver/store/service/StoreService.java
@@ -1,22 +1,22 @@
-package com.sparta.blackwhitedeliverydriver.service;
+package com.sparta.blackwhitedeliverydriver.store.service;
 
-import com.sparta.blackwhitedeliverydriver.dto.StoreIdResponseDto;
-import com.sparta.blackwhitedeliverydriver.dto.StoreRequestDto;
-import com.sparta.blackwhitedeliverydriver.dto.StoreResponseDto;
-import com.sparta.blackwhitedeliverydriver.dto.StoreByMasterRequestDto;
-import com.sparta.blackwhitedeliverydriver.entity.Category;
-import com.sparta.blackwhitedeliverydriver.entity.Store;
-import com.sparta.blackwhitedeliverydriver.entity.StoreCategory;
-import com.sparta.blackwhitedeliverydriver.entity.User;
-import com.sparta.blackwhitedeliverydriver.entity.UserRoleEnum;
-import com.sparta.blackwhitedeliverydriver.exception.CategoryExceptionMessage;
-import com.sparta.blackwhitedeliverydriver.exception.ExceptionMessage;
-import com.sparta.blackwhitedeliverydriver.exception.StoreExceptionMessage;
-import com.sparta.blackwhitedeliverydriver.repository.CategoryRepository;
-import com.sparta.blackwhitedeliverydriver.repository.StoreCategoryRepository;
-import com.sparta.blackwhitedeliverydriver.repository.StoreRepository;
-import com.sparta.blackwhitedeliverydriver.repository.UserRepository;
-import com.sparta.blackwhitedeliverydriver.security.UserDetailsImpl;
+import com.sparta.blackwhitedeliverydriver.store.dto.StoreIdResponseDto;
+import com.sparta.blackwhitedeliverydriver.store.dto.StoreRequestDto;
+import com.sparta.blackwhitedeliverydriver.store.dto.StoreResponseDto;
+import com.sparta.blackwhitedeliverydriver.store.dto.StoreByMasterRequestDto;
+import com.sparta.blackwhitedeliverydriver.category.entity.Category;
+import com.sparta.blackwhitedeliverydriver.store.entity.Store;
+import com.sparta.blackwhitedeliverydriver.store.entity.StoreCategory;
+import com.sparta.blackwhitedeliverydriver.user.entity.User;
+import com.sparta.blackwhitedeliverydriver.user.entity.UserRoleEnum;
+import com.sparta.blackwhitedeliverydriver.category.exception.CategoryExceptionMessage;
+import com.sparta.blackwhitedeliverydriver.user.exception.ExceptionMessage;
+import com.sparta.blackwhitedeliverydriver.store.exception.StoreExceptionMessage;
+import com.sparta.blackwhitedeliverydriver.category.repository.CategoryRepository;
+import com.sparta.blackwhitedeliverydriver.store.repository.StoreCategoryRepository;
+import com.sparta.blackwhitedeliverydriver.store.repository.StoreRepository;
+import com.sparta.blackwhitedeliverydriver.user.repository.UserRepository;
+import com.sparta.blackwhitedeliverydriver.common.security.UserDetailsImpl;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotBlank;
 import java.time.LocalDateTime;
